Refactor feature packages & remove config
