--- src/main/java/net/countercraft/movecraft/combat/config/Config.java
@@ -1,6 +1,0 @@
-package net.countercraft.movecraft.combat.config;
-
-@Deprecated(forRemoval = true)
-public class Config {
-    // Combat Releasing
-}
