+++ src/main/java/net/countercraft/movecraft/combat/features/combat/events/CombatEvent.java
@@ -1,4 +1,4 @@
-package net.countercraft.movecraft.combat.event;
+package net.countercraft.movecraft.combat.features.combat.events;
 
 import org.jetbrains.annotations.Nullable;
 import org.bukkit.entity.Player;
