+++ src/main/java/net/countercraft/movecraft/combat/features/combat/events/CombatReleaseEvent.java
@@ -1,4 +1,4 @@
-package net.countercraft.movecraft.combat.event;
+package net.countercraft.movecraft.combat.features.combat.events;
 
 import org.bukkit.entity.Player;
 import org.bukkit.event.Cancellable;
