+++ src/main/java/net/countercraft/movecraft/combat/features/tracking/TNTTracking.java
@@ -1,9 +1,9 @@
-package net.countercraft.movecraft.combat.features.damagetracking;
+package net.countercraft.movecraft.combat.features.tracking;
 
 import net.countercraft.movecraft.combat.MovecraftCombat;
 import net.countercraft.movecraft.combat.features.CannonDirectors;
-import net.countercraft.movecraft.combat.features.CombatRelease;
-import net.countercraft.movecraft.combat.features.damagetracking.types.TNTCannonDamage;
+import net.countercraft.movecraft.combat.features.combat.CombatRelease;
+import net.countercraft.movecraft.combat.features.tracking.types.TNTCannonDamage;
 import net.countercraft.movecraft.craft.Craft;
 import net.countercraft.movecraft.craft.PlayerCraft;
 import org.bukkit.entity.Player;
