+++ src/main/java/net/countercraft/movecraft/combat/features/tracking/types/DamageType.java
@@ -1,4 +1,4 @@
-package net.countercraft.movecraft.combat.features.damagetracking.types;
+package net.countercraft.movecraft.combat.features.tracking.types;
 
 
 // Interface for all types of damage
