+++ src/main/java/net/countercraft/movecraft/combat/features/tracking/types/FireballDamage.java
@@ -1,4 +1,4 @@
-package net.countercraft.movecraft.combat.features.damagetracking.types;
+package net.countercraft.movecraft.combat.features.tracking.types;
 
 public class FireballDamage implements DamageType {
     public String toString() {
