+++ src/main/java/net/countercraft/movecraft/combat/listener/ProjectileLaunchListener.java
@@ -1,8 +1,8 @@
 package net.countercraft.movecraft.combat.listener;
 
-import net.countercraft.movecraft.combat.features.CombatRelease;
-import net.countercraft.movecraft.combat.features.damagetracking.DamageTracking;
-import net.countercraft.movecraft.combat.features.damagetracking.FireballTracking;
+import net.countercraft.movecraft.combat.features.combat.CombatRelease;
+import net.countercraft.movecraft.combat.features.tracking.DamageTracking;
+import net.countercraft.movecraft.combat.features.tracking.FireballTracking;
 import net.countercraft.movecraft.craft.Craft;
 import net.countercraft.movecraft.craft.CraftManager;
 import net.countercraft.movecraft.craft.PlayerCraft;
